Fix packages declaration of some tests.
Tests are now in same package as tested class.
(See larsga/Duke#150 for more information)
