+++ duke-core/src/test/java/no/priv/garshol/duke/cleaners/DigitsOnlyCleanerTest.java
@@ -1,12 +1,11 @@
 
-package no.priv.garshol.duke.test;
+package no.priv.garshol.duke.cleaners;
 
-import org.junit.Test;
 import org.junit.Before;
-import static junit.framework.Assert.assertTrue;
-import static junit.framework.Assert.assertEquals;
+import org.junit.Test;
 
-import no.priv.garshol.duke.cleaners.DigitsOnlyCleaner;
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertTrue;
 
 public class DigitsOnlyCleanerTest {
   private DigitsOnlyCleaner cleaner;
