+++ duke-core/src/test/java/no/priv/garshol/duke/cleaners/NorwegianCompanyNameCleanerTest.java
@@ -1,11 +1,10 @@
 
-package no.priv.garshol.duke.test;
+package no.priv.garshol.duke.cleaners;
 
-import org.junit.Test;
 import org.junit.Before;
-import static junit.framework.Assert.assertEquals;
+import org.junit.Test;
 
-import no.priv.garshol.duke.cleaners.NorwegianCompanyNameCleaner;
+import static junit.framework.Assert.assertEquals;
 
 public class NorwegianCompanyNameCleanerTest {
   private NorwegianCompanyNameCleaner cleaner;
