+++ duke-core/src/test/java/no/priv/garshol/duke/comparators/MetaphoneComparatorTest.java
@@ -1,11 +1,9 @@
 
-package no.priv.garshol.duke.test;
+package no.priv.garshol.duke.comparators;
 
 import org.junit.Test;
-import static junit.framework.Assert.assertTrue;
-import static junit.framework.Assert.assertEquals;
 
-import no.priv.garshol.duke.comparators.MetaphoneComparator;
+import static junit.framework.Assert.assertEquals;
 
 public class MetaphoneComparatorTest {
