+++ duke-core/src/test/java/no/priv/garshol/duke/comparators/NorphoneComparatorTest.java
@@ -1,12 +1,10 @@
 
-package no.priv.garshol.duke.test;
+package no.priv.garshol.duke.comparators;
 
 import org.junit.Test;
-import static junit.framework.Assert.assertTrue;
+
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.fail;
-
-import no.priv.garshol.duke.comparators.NorphoneComparator;
 
 public class NorphoneComparatorTest {
