+++ duke-core/src/test/java/no/priv/garshol/duke/comparators/QGramComparatorTest.java
@@ -1,14 +1,10 @@
 
-package no.priv.garshol.duke.test;
+package no.priv.garshol.duke.comparators;
 
-import org.junit.Test;
 import org.junit.Before;
-import static junit.framework.Assert.fail;
-import static junit.framework.Assert.assertTrue;
-import static junit.framework.Assert.assertEquals;
+import org.junit.Test;
 
-import no.priv.garshol.duke.DukeConfigException;
-import no.priv.garshol.duke.comparators.QGramComparator;
+import static junit.framework.Assert.assertEquals;
 
 public class QGramComparatorTest {
   private QGramComparator comp;
