+++ duke-core/src/test/java/no/priv/garshol/duke/databases/InMemoryBlockingDatabaseTest.java
@@ -1,14 +1,12 @@
 
-package no.priv.garshol.duke.test;
+package no.priv.garshol.duke.databases;
 
 import java.util.ArrayList;
 import java.util.Collection;
 
-import no.priv.garshol.duke.Record;
-import no.priv.garshol.duke.Database;
 import no.priv.garshol.duke.Configuration;
-import no.priv.garshol.duke.databases.KeyFunction;
-import no.priv.garshol.duke.databases.InMemoryBlockingDatabase;
+import no.priv.garshol.duke.Database;
+import no.priv.garshol.duke.Record;
 
 public class InMemoryBlockingDatabaseTest extends DatabaseTest {
