+++ duke-core/src/test/java/no/priv/garshol/duke/test/DeduplicatorTest.java
@@ -1,30 +1,27 @@
 
 package no.priv.garshol.duke.test;
 
-import org.junit.Test;
-import org.junit.After;
-import org.junit.Before;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.assertEquals;
-
-import java.util.List;
-import java.util.Iterator;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.io.IOException;
+import java.util.Iterator;
+import java.util.List;
 
-import no.priv.garshol.duke.Record;
-import no.priv.garshol.duke.Property;
-import no.priv.garshol.duke.Processor;
-import no.priv.garshol.duke.PropertyImpl;
-import no.priv.garshol.duke.RecordIterator;
 import no.priv.garshol.duke.ConfigurationImpl;
+import no.priv.garshol.duke.Processor;
+import no.priv.garshol.duke.Property;
+import no.priv.garshol.duke.PropertyImpl;
+import no.priv.garshol.duke.Record;
+import no.priv.garshol.duke.RecordIterator;
 import no.priv.garshol.duke.comparators.Levenshtein;
-import no.priv.garshol.duke.utils.DefaultRecordIterator;
-import no.priv.garshol.duke.matchers.AbstractMatchListener;
-import no.priv.garshol.duke.matchers.PrintMatchListener;
 import no.priv.garshol.duke.datasources.InMemoryDataSource;
+import no.priv.garshol.duke.utils.DefaultRecordIterator;
+import no.priv.garshol.duke.utils.TestUtils;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
 
 public class DeduplicatorTest {
   private ConfigurationImpl config;
